imx8m-var-dart: bcm43xx-utils; Reworked WIFI start-up scripts
Modified start-up sequence to avoid 5s delay on boot
Moved common functions from variscite-wifi to variscite-wifi-common.sh

+++ recipes-connectivity/bcm43xx-utils/bcm43xx-utils/variscite-wifi-common.sh
@@ -1,6 +1,55 @@
+# Setup WIFI control GPIOs
+wifi_setup()
+{
+	if [ ! -d /sys/class/gpio/gpio${WIFI_VSEL_GPIO} ]; then
+		echo ${WIFI_VSEL_GPIO} > /sys/class/gpio/export
+		echo out > /sys/class/gpio/gpio${WIFI_VSEL_GPIO}/direction
+	fi
+
+	if [ ! -d /sys/class/gpio/gpio${WIFI_PWR_GPIO} ]; then
+		echo ${WIFI_PWR_GPIO} > /sys/class/gpio/export
+		echo out > /sys/class/gpio/gpio${WIFI_PWR_GPIO}/direction
+	fi
+
+	if [ ! -d /sys/class/gpio/gpio${WIFI_EN_GPIO} ]; then
+		echo ${WIFI_EN_GPIO} > /sys/class/gpio/export
+		echo out > /sys/class/gpio/gpio${WIFI_EN_GPIO}/direction
+	fi
+
+	if [ ! -d /sys/class/gpio/gpio${BT_BUF_GPIO} ]; then
+		echo ${BT_BUF_GPIO} > /sys/class/gpio/export
+		echo out > /sys/class/gpio/gpio${BT_BUF_GPIO}/direction
+	fi
+
+	if [ ! -d /sys/class/gpio/gpio${BT_EN_GPIO} ]; then
+		echo ${BT_EN_GPIO} > /sys/class/gpio/export
+		echo out > /sys/class/gpio/gpio${BT_EN_GPIO}/direction
+	fi
+
+	# WIFI_VSEL up
+	echo 1 > /sys/class/gpio/gpio${WIFI_VSEL_GPIO}/value
+	usleep 10000
+}
+
+# Check that WIFI interface exists
+wifi_is_up()
+{
+	for i in $(seq 1 20); do
+		[ -d /sys/class/net/wlan0 ] && return 0
+		sleep 1
+	done
+
+	return 1
+}
+
 # Power up WIFI chip
 wifi_up()
 {
+	# Unbind WIFI device from MMC controller
+	if [ -e /sys/bus/platform/drivers/sdhci-esdhc-imx/${WIFI_MMC_HOST} ]; then
+	   echo ${WIFI_MMC_HOST} > /sys/bus/platform/drivers/sdhci-esdhc-imx/unbind
+        fi
+
 	# WIFI_PWR up
 	echo 1 > /sys/class/gpio/gpio${WIFI_PWR_GPIO}/value
 	usleep 10000
@@ -14,7 +63,7 @@ wifi_up()
 	# BT_BUF up
 	echo 0 > /sys/class/gpio/gpio${BT_BUF_GPIO}/value
 	
-	# Wait 150ms at least
+	# Wait at least 150ms
 	usleep 200000
 	
 	# BT_BUF down
@@ -43,7 +92,9 @@ wifi_down()
 	modprobe -r fec
 
 	# Unbind WIFI device from MMC controller
-	echo ${WIFI_MMC_HOST} > /sys/bus/platform/drivers/sdhci-esdhc-imx/unbind
+	if [ -e /sys/bus/platform/drivers/sdhci-esdhc-imx/${WIFI_MMC_HOST} ]; then
+	   echo ${WIFI_MMC_HOST} > /sys/bus/platform/drivers/sdhci-esdhc-imx/unbind
+	fi
 
 	# WLAN_EN down
 	echo 0 > /sys/class/gpio/gpio${WIFI_EN_GPIO}/value
@@ -53,7 +104,6 @@ wifi_down()
 
 	# BT_EN down
 	echo 0 > /sys/class/gpio/gpio${BT_EN_GPIO}/value
-
 	usleep 10000
 
 	# WIFI power down
